test_gemini.py
import os
from dotenv import load_dotenv
import google.generativeai as genai

load_dotenv()

genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

print("Testing Gemini access...\n")

try:
    models = genai.list_models()
    if not models:
        print("No models returned")
    for m in models:
        print(m.name, m.supported_generation_methods)
except Exception as e:
    print("ERROR:", e)
