requirements.txt
streamlit==1.32.0
google-generativeai==0.4.1
yfinance==0.2.33
python-dotenv==1.0.1
langgraph==0.0.38
